# docs/swagger_simple.yaml
openapi: 3.0.0
info:
  title: Sports Academy Management API
  description: API for managing sports academy operations
  version: 1.0.0
  contact:
    email: info@sportsacademy.com
servers:
  - url: http://localhost:5000/api/v1
    description: Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /tournaments/{id}:
    get:
      tags:
        - Tournaments
      summary: Get a specific tournament
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Tournament ID
      responses:
        '200':
          description: Tournament details
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: object

    patch:
      tags:
        - Tournaments
      summary: Update tournament details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Tournament ID
      responses:
        '200':
          description: Tournament updated

    delete:
      tags:
        - Tournaments
      summary: Delete tournament
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Tournament ID
      responses:
        '204':
          description: Tournament deleted
...